requirements.txt
pandas==2.2.3
numpy==1.25.2
matplotlib==3.8.1
seaborn==0.12.2
tqdm==4.65.0

scikit-learn==1.3.2
pyod==1.1.0
prophet==1.1.4

streamlit==1.24.1
gradio==3.40.1

pdfplumber==0.7.6
camelot-py[cv]==0.10.1
tabula-py==2.8.0

transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.2.2

pyttsx3==2.90
gTTS==2.3.1

python-dotenv==1.0.0
shap==0.41.0
